fix(user): fix start & end meeting time synchronization
